Allocator::guaranteedAlignment(), add common memory size literals, use uint64 over size_t

+++ src/filesystem/FileSystem.hpp
@@ -53,7 +53,7 @@ public:
 	}
 	
 	template <typename T>
-	void readValueArray(T* array, size_t N) {
+	void readValueArray(T* array, int64 N) {
 		readBytes(array, sizeof(T) * N);
 	}
